+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
-## [1.4.0] - Unreleased
+## [2.0.0] - Unreleased
 ### Added
 - latestN() function.
+- get() function.
+- value() function.
+- refper() function.
+- length() function.
 - Support for Internet Explorer 11.
 
 ### Changed